Moved the zarr chunk stats code to universal_utilities.py and ran again. It worked fine. Now I can apply it to the main model.
